+++ src/__tests__/model-cache.test.ts
@@ -7,7 +7,7 @@ describe('ModelCache', () => {
   beforeEach(() => {
     vi.unstubAllEnvs();
     cache = ModelCache.getInstance();
-    cache.setModels([]);
+    cache.reset();
   });
 
   afterEach(() => {